+++ CHANGELOG.md
@@ -5,10 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/) and this
 project adheres to [Semantic Versioning](http://semver.org/).
 
-## Unreleased
+<!-- ## Unreleased -->
+
+## [0.5.2] 2020-09-08
 
 - Add support for `name` property to measurement config
 - Include `measurement` object for each benchmark in JSON results output
+- Fix errors with auto-installing WebDriver modules using Yarn
+  ([#186](https://github.com/Polymer/tachometer/issues/186])).
 
 ## [0.5.1] 2020-08-19